FIX: Preventing a single grounding to be added twice.
If two users simultaneously add the same grounding, it is added only once.

More than other memos, graphs are likely to be intensively edited by different users. Consecutively, the list of existing nodes is not any more compiled on the client side when the graph is refreshed; instead the database is consulted when an addition is submitted.