REFACTORING: Paths to memos are computed on the server side.
`list/diary.js` computes the relative URLs to memos, diagrams, tables and graphs, which allows for a faster diary page rendering.

The list of memos and the list of candidates for deletion are now integrated, which lightens the diary page and allows for the list to be shown as soon as the page is rendered (#75).