REFACTORING: Introducing field memos (distinct from transcripts).
The original design regards field memo as transcripts. To bypass this limitation, users thus entered field notes as the first speech turn of an interview, which is not convenient.

Experience definitely advocates the creation of an additional type of memos, dedicated to observation notes. Such a memo behaves just as operational, coding, theoretical or storyline memos. Among other features, it thus allows to insert pictures (#68).