REFACTORING: Creating memos.
The input field and button allowing to create a new memo, diagram, table or graph are moved in the footer navbar. This avoids the input field to cover the title (on medium and small screens). This new location is also coherent with the separation of browsing functions (in the header navbar) and edit functions (in the footer navbar).

This commit also contains minor fixes :
* Diary and memo header collapses on medium screens.
* The groundings panel stays on the left of the screen.
* Memos previews show only when mouse hovers their name.