FIX: When no reader is mentioned, everyone can read.
Up to now, documents (memos, diagrams, tables and graphs) with no `readers` field behaved differently from those with an empty `readers` list. This did not raise any issue, given it was impossible for users to remove readers themselves. Now that this feature is provided, the situation will occur. This fix prevents unconsistent behaviour in this matter.