REFACTORING: An alert notifies of pending tasks (#65).
The "sticky" tasklist icon (on the right side of the diary page) proved to be unnoticed. It is thus replaced by a (more noticeable) alert. The icon is kept on the right side of the alert, to prevent experienced users to be confused.

The color of the (updated/total) memos counts (right to the diary name) is harmonized with the highlights used on the diaries page. Next to this count, number of unchecked comments is provided in another color.