FEATURE: Maintenance/Downtime notification.
This notification feature helps the server admin to warn the users (with no need to send emails). Downtime moment is scheduled in a dedicated document (named `maintenance`). Notifications are shown (as Bootstrap toasts) in the groundings area of every memo on the platform. Even unregistered users can be warned this way, which prevents inconvenience for every user.