FEATURE: Memos can be edited by one person at once.
When a memo is edited, the "edit" button is disabled (except for the user that started the edit, who is allowed to resume the edition at any time).
When the button is disabled, a tooltip informs the user that the memo is edited (mentioning when the edition started and who started it).

Edition mode replaces "draft" mode (reserved to transcript) :
* Lexicometrics computation is performed when edition ends (it is not during edition).
* Tasklist mentions memos that are edited.