REFACTORING: On the diary page, memos are incementally listed (#75).
The diary page does not list all belonging memos at each load. Instead, only 30 memos are listed (depending on the sorting order chosen by the user). The following memos are listed if the user scrolls down the page. The number of memos composing the diary and the number of unreads memos are thus processed before memos are listed.

Timeline and network views will be computed on different pages.

This commit includes fixes to the stringified function used in the rewriting process.