FEATURE: Sorting order saved as user preference (#89).
When the user selects how memos should be sorted on the diary page, the selected order is saved as a preference, so that the list will be sorted in the chosen order, even if the user closes the session. Preferences are stored on the server side (so that it is kept, even if the user connects from a different computer).

This commit also fixes how "Relative Moments" function is mutualised.