FEATURE: Grounding an existing memo in the current memo (#83)
Up to now, if the user tried to create a memo with a name that already existed in the diary, he was prompted (though a dialog box) that a memo with that name already exists.

From now on, the platform suggests to ground the existing memo in the current one (through a one-click operation). The user is simply prompted (with no suggestion) if the grounding link already exists.