REFACTORING: "Restricted access" becomes "Authentication required"
Users may be confused to be prompted that they have no access to their memos after their authentication expired. To avoid such a confusion, the "Restricted access" message becomes "Authentication required".

In the future, it would be convenient to provide two different messages for unlogged users and for logged users trying to access a reserved memo.