REFACTORING: Edit button appears when the memo is ready
When a memo is being edited by another user, the memo is locked and cannot be edited. However, given the "edit" button is hidden at the end of the loading process, a user with slow Internet connexion has access to the edit button (even when s/he should not). Under such circumstances, the user may edit memo and lose data. To prevent such a situation, edit buttons are hidden by default and uncovered at the end of the loading process.