REFACTORING: Preventing new transcript to be created
We know that the transcript editing form contains some bugs: when editing, speech turns may duplicate or disappear. We have been able to reproduce the bugs but unable to understand their cause. Therefore, the problem was not fixed.

To prevent the problem to occur, the button allowing to create transcripts is removed.
* New transcripts will be included as field memos
* The template used to show transcripts and the edit form are kept, in order for the already existing transcripts to work as before.

In the future, the edit form may be modified in a way that convert the edited transcript into field memo.