REFACTORING: Situating empirical material
When the user creates a field memo, s/he is prompted with a dialog requesting the date and the location of field observations and, in the case of an interview transcript, the pseudonym of the informant. These attributes are used to entitle the memo.

This dialog reinstates the `attributes` form that @benel designed for the first version of Cassandre.