FIX: Preventing double conversion of interviews (#104)
Some users visit `editable_text` URLs thanks to the list of pages previously visited with their browser. This unexpected behaviour triggers the conversion script, sometimes for memos already converted into interview type. In such a case, the resulting memo is deprived of the `diary` field. To avoid this undesired situation, the conversion occurs only if no `body` field is available.