FEATURE: Diagram name auto-updates
When one user renames one of the properties belonging to a diagram, the name of the diagram is updated. However, this process happens to fail on some (unidentified) occasions. To fix this problem, the diagram name will automatically updates when one user (with rigths to modify the diagram) visits the diagram.