REFACTORING: Activity follows access rights
Instead of showing all activities independently from user rights, the activity page henceforth shows creations, modifications and comments of memos that the user is allowed to access. This feature prevents to show all activities (and, incidentally all memo names) of a diary to a user who is allowed to read only one memo of this diary.

This refactoring is done at list level; as #98 suggests for task-list, a deeper (view level) refactoring could be more efficient.