FEATURE: Gathering statements
When one user has putted articulations into words (a statement) under a least one diagram, a memo appears among concerned diagrams leaves. The body of this special memo gathers all formulated statements. The groundings of this special memo refer to the diagrams that provides the statements. Its leaves consist of all graphs created in the diary.

This special memo lists statements from diagrams that the user has the right to read. If the user is allowed to read some (but not all) diagrams, only the one s/he is allowed to read are featured.

This special memo cannot be directly edited; its content, groundings and leaves are automatically updated according to diagrams and graphs. This page automatically refreshes when one of its grounding(s) is modified.

The page also provides a field to create a storyline memo. If such a storyline memo is created from this special memo, it rests on all diagrams (i.e. the groundings are the same as the special memo) and its (read/modify) rights aggregate those from the grounding diagrams.

This feature celebrates the 500th commit (#105)