REFACTORING: Preventing memo locking
During its edition, a memo cannot be modified by other users. When leaving a memo unsaved, unexperienced users hinder teamwork. From now on, a more explicit message (in a toast) helps the user to be aware of the consequences of leaving a memo unsaved (#99).

Locked memos also result from visiting the editing form (thanks to the back button or through browser history). Such visits unwillingly change the `editing` status of the memo. To prevent this, the `editing` flag is now triggered by the `edit` button. Loading the editing form by accident will not trigger this editing state any more. Moreover, access is prohibited if the editing form is requested for a memo with no `editing` flag. It is replaced by an explicit message discouraging using the back button.