FEATURE: A reader can unsubscribe from a memo (#101)
Readers can read but not modify the memo (including access rights). As a consequence, a reader cannot unsubscribe from a memo. To allow a reader to unsubscribe from a memo, a special exception was created in the validation function. Thanks to this exception, a dedicated update function allows readers to unsubscribe. To prevent the memo to become public, this feature replaces the last reader that unsubscribe with the user that created the memo of the first user in the contributors list.