REFACTORING: Preventing unintended memo edition
Although not intending to edit a memo, some users visit the editing form (thanks to the back button or through browser history). Since commit 8c695cbe1e1fc11193060fd490de0114f70349bb , an explicit message asked the user to avoid using the back button. Up to now, this message is removed and, when no `editing` flag is on, the editing form instantly redirects the user to the memo page. This behaviour is less disconcerting and matches better what users intend.