FIX: Updating diagrams name
When the name of one articulated dimension/property is edited, the name 
of the diagram representing the articulation is updated. 

In order for this name to be correctly updated, the two parts of the 
diagram name are henceforth ordered according to the articulated coding 
memo. 

The id of its (two) groundings are thus stored in alphabetical order. 
This commit includes a fix for already created diagrams (whose 
groundings order is determined by which one is articulated first).