FIX: Diagram name auto-updates
The diagram name automatically updates when one user (with rights to 
modify the diagram) visits the diagram 
(d20c6c5312eb1d38e0919906792df2ca38183cc8). This update should not be 
triggered because of tabs or newlines in the names of articulated 
dimensions/properties. The following fixes prevent such undesired 
updates.

* The renaming process escapes tabs and newlines.
* Tabs and newlines are excluded from memo names (when they are created 
or edited). 
* Automatic renaming is not recorded in user activity. 
* Articulation cannot have more than two groundings.