FEATURE: Highlighting searched or coded paragraphs
When some text is coded or searched, the result is provided with a link 
pointing to the field memo where the text comes from. When such a link 
is followed, the field memo is opened and scrolled to the matching 
paragraph, that is highlighted in yellow for a few seconds.

This (long-awaited) feature pursues the direction opened with  
948d2af0e836e013544bddc8459cb34e0c48754c (#106).