FEATURE: Prompting the user for full name
When an external/institutional LDAP is used (thanks to AAAforREST), 
users may log in without prior registering. In that case, the platform 
does not know the user full name. When this happens, the user is 
prompted for her full name. The user document is created at this stage, 
so that the personal diary order and history can be stored.