REFACTORING: Lazy loading for memos and diagrams
Memos and diagrams are loaded in two steps. The document itself is loaded with a show. Comments, groundings, leaves, users fullnames and diary name are loaded in a second time,  thanks to a list (named satellites). When the server is under heavy load, the memo content can be shown prior to `satellites` details.