FEATURE: Grounding an existing memo
When the user fills the add/create input field, a list of existing memos is suggested (as #83 proposed), allowing to add the current memo among the groundings of the selected memo.

If the user requests an already existing grounding, s/he is prompted that it already exists. If the memo is already grounded, but on a different fragment/paragraph, the grounding is quietly updated.

This feature can contribute to smoother coding sessions, as @benel indicates (#112). That's why it is reserved to field memos and transcripts at this stage.