FEATURE: Each memo page mentions his/her creator (#127)
The creation date and the fullname of the creator move from the "revert" tootip to a dedicate tag with the initials of the creator, in the header of each memo, diagram or graph. As a consequence, this information becomes available to readers; it is not reserved to contributors any more.