FEATURE: Modify rights at diary level
The right to read or to modify memos can be granted or removed at once for all memos belonging to one diary. Of course, only edtiable memos are affected by this feature. As a consequence, even if requested for a complete diary, such a request produces different results according to the rights of the requesting user.

To prevent the accidental publication of memo (ie memo world-wide readable), the removal of the last reader should be accomplished in two passes. Thefirst pass replaces the last user by the user requesting the removal. If the last user is the current user, the removal is done, and public access is granted world-wide.