FEATURE: Users can remove or modify their avatar image
When no avatar is selected or when it is removed, the creator tag shows colored initials (as before).

For performance, avatar image names are locally stored (as fullnames). This local storage is updated when the user visits a memo created by the person whose avatar is modified. Creator tags are thus not instantly updated.